% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_archive}
\alias{data_archive}
\title{Move data to archive for a state (equal-to or before a period)}
\usage{
data_archive(
  state,
  period,
  current_drive = file.path("E:/SA/Data-sensitive/Data-Dashboards", state),
  archive_drive = file.path("H:/SA/Data-sensitive/Data-Dashboards", state),
  pattern = "raw-2...-q.",
  move_folders = TRUE,
  move_files = FALSE
)
}
\arguments{
\item{state}{character: Two letter state designation}

\item{period}{character: Time period for the first project 
(e.g., "2019-q4", "2020-q4", etc.)}

\item{current_drive}{drive from which to move}

\item{archive_drive}{destination drive}

\item{pattern}{matching pattern sent to \code{\link[base]{list.files}}}

\item{move_folders}{If TRUE, will move files within folders matching the pattern}

\item{move_files}{If TRUE, will move files matching the pattern (e.g., sqlite)}
}
\description{
Probably most useful for moving in batches (e.g., looping across states). 
Othewise it can be done easily by hand. By default, moves relevant raw data 
files to the H drive.
}
\seealso{
Other functions to manage data: 
\code{\link{data_destroy}()}
}
\concept{functions to manage data}
