% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{data_destroy}
\alias{data_destroy}
\title{(Not implemented) Permanently remove data}
\usage{
data_destroy(
  state,
  period,
  drive = file.path("H:/SA/Data-sensitive/Data-Dashboards", state),
  force = FALSE
)
}
\arguments{
\item{state}{character: Two letter state designation}

\item{period}{character: Time period for the first project 
(e.g., "2019-q4", "2020-q4", etc.)}

\item{drive}{target location from which to remove files}

\item{force}{If TRUE, delete without prompt (potentially useful for
batch deletes). Use with caution.}
}
\description{
By default you will be prompted to enter "yes" to confirm that you want to
remove a specified set of files.
}
\seealso{
Other functions to update projects: 
\code{\link{data_archive}()},
\code{\link{setup_data_dive}()},
\code{\link{update_project}()}
}
\concept{functions to update projects}
