% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_periods}
\alias{get_periods}
\title{Helper function to identify files with equal to or lower given period}
\usage{
get_periods(files, period)
}
\arguments{
\item{files}{vector of files/folders}

\item{period}{character: Time period for the first project 
(e.g., "2019-q4", "2020-q4", etc.)}
}
\description{
To be called from \code{\link{data_archive}}
}
\examples{
files <- c("raw-2016-q2", "raw-2016-q4", "raw-2016-q4.sqlite3",
           "raw-2017-q2", "raw-2017-q4")
get_periods(files, "2016-q4")
}
\seealso{
Other helper functions: 
\code{\link{replace_params}()},
\code{\link{setup_project}()}
}
\concept{helper functions}
