% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{new_project}
\alias{new_project}
\title{Setup a new state project with default directories and template scripts}
\usage{
new_project(
  state,
  period,
  analysis_path = file.path("E:/SA/Projects/Data-Dashboards", state, period),
  sensitive_path = file.path("E:/SA/Data-Sensitive/Data-Dashboards", state,
    paste0("raw-", period)),
  production_path = file.path("E:/SA/Data-Production/Data-Dashboards", state)
)
}
\arguments{
\item{state}{character: Two letter state designation}

\item{period}{character: Time period for the first project 
(e.g., "2019-q4", "2020-q4", etc.)}

\item{analysis_path}{character: File path for the analysis}

\item{sensitive_path}{character: File path for the sensitive data. If NULL,
no folders will be created.}

\item{production_path}{character: File path for the production data. If NULL, 
no folders will be created.}
}
\description{
This is intended to be run before data processing for a new state begins.
}
\examples{
# new_project("YY", "2019-q4")
}
\concept{functions for making template files/folders}
