% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{new_project_individual}
\alias{new_project_individual}
\title{Setup a new individual state dashboard with default directories and template scripts}
\usage{
new_project_individual(
  state,
  period,
  template_path = "template-individual",
  print_message = "A new individual state dashboard has been initialized",
  ...
)
}
\arguments{
\item{state}{character: Two letter state designation}

\item{period}{character: Time period for the first project 
(e.g., "2019-q4", "2020-q4", etc.)}

\item{template_path}{character: Name of inst subpath in the package that points
to the template files}

\item{print_message}{character: Message to print on success}

\item{...}{other arguments passed to \code{\link{new_project}}}
}
\description{
For states which have custom individual dashboards (in addition to national/regional).
This function is intended to be run before data processing for a new state begins. 
It's just a wrapper for \code{\link{new_project}}.
}
\examples{
# new_project_individual("YY", "2019-q4")
}
\seealso{
Other functions for new projects based on templates: 
\code{\link{new_project}()}
}
\concept{functions for new projects based on templates}
