% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{new_project_summary}
\alias{new_project_summary}
\title{Setup a new summary state dashboard with default directories and template scripts}
\usage{
new_project_summary(
  state,
  period,
  template_path = "template-summary",
  print_message = "A new summary data dashboard project has been initialized",
  ...
)
}
\arguments{
\item{state}{character: Two letter state designation}

\item{period}{character: Time period for the first project 
(e.g., "2019-q4", "2020-q4", etc.)}

\item{template_path}{character: Name of inst subpath in the package that points
to the template files}

\item{print_message}{character: Message to print on success}

\item{...}{other arguments passed to \code{\link{new_project}}}
}
\description{
For states which process their own dashboard summaries.
}
\examples{
# new_project_summary("YY", "2019-q4")
}
\seealso{
Other functions for new projects based on templates: 
\code{\link{new_project_individual}()},
\code{\link{new_project}()}
}
\concept{functions for new projects based on templates}
