% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{setup_data_dive}
\alias{setup_data_dive}
\title{Setup data dive template}
\usage{
setup_data_dive(
  state,
  period,
  analysis_path = file.path("E:/SA/Projects/Data-Dashboards", state, period),
  dive_path = file.path(analysis_path, "code", "5-data-dive")
)
}
\arguments{
\item{state}{character: Two letter state designation}

\item{period}{character: Time period for the first project 
(e.g., "2019-q4", "2020-q4", etc.)}

\item{analysis_path}{character: File path for the analysis}

\item{dive_path}{folder for data dive code}
}
\description{
This is to be run for an existing state-time_period dashboard project. 
It creates a new folder with template code. Defaults to placing a "5-data-dive"
folder in the working directory (i.e., when state and time_period are NULL).
}
\examples{
# setup_data_dive("YY", "2019-q4")
}
\seealso{
Other functions to update projects: 
\code{\link{update_project}()}
}
\concept{functions to update projects}
