% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{setup_project}
\alias{setup_project}
\title{Helper function for setting up project code}
\usage{
setup_project(state, period, analysis_path, sensitive_path, production_path)
}
\arguments{
\item{state}{character: Two letter state designation}

\item{period}{character: Time period for the first project 
(e.g., "2019-q4", "2020-q4", etc.)}

\item{analysis_path}{character: File path for the analysis}

\item{sensitive_path}{character: File path for the sensitive data. If NULL,
no folders will be created.}

\item{production_path}{character: File path for the production data. If NULL, 
no folders will be created.}
}
\description{
Only intended to be run from other functions: (\code{\link{new_project}}), etc.
It does some error checking and sets up analysis/data folders.
}
\seealso{
Other helper functions: 
\code{\link{get_periods}()},
\code{\link{replace_params}()}
}
\concept{helper functions}
