% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{update_project}
\alias{update_project}
\title{Setup a new period for a project}
\usage{
update_project(
  state,
  period,
  ref_period,
  ref_path = file.path("E:/SA/Projects/Data-Dashboards", state, ref_period),
  analysis_path = file.path("E:/SA/Projects/Data-Dashboards", state, period),
  sensitive_path = file.path("E:/SA/Data-Sensitive/Data-Dashboards", state,
    paste0("raw-", period)),
  production_path = file.path("E:/SA/Data-Production/Data-Dashboards", state),
  files_to_keep = c("\\\\.r", "documentation\\\\.tex", "\\\\.txt"),
  top_level_files = c(".Rprofile", "README.md", "renv.lock", "renv/activate.R",
    "renv/settings.dcf")
)
}
\arguments{
\item{state}{character: Two letter state designation}

\item{period}{character: Time period for the first project 
(e.g., "2019-q4", "2020-q4", etc.)}

\item{ref_period}{folder name of reference time period (e.g., 2019-q4)}

\item{ref_path}{full file path to reference period analysis folder}

\item{analysis_path}{character: File path for the analysis}

\item{sensitive_path}{character: File path for the sensitive data. If NULL,
no folders will be created.}

\item{production_path}{character: File path for the production data. If NULL, 
no folders will be created.}

\item{files_to_keep}{patterns for matching on files in the "code" folder using grep 
(not case sensitive). Files with matching patterns will be copied from ref_period.}

\item{top_level_files}{top-level analysis files that should be copied over}
}
\description{
This will setup files based on a reference time period, rather than 
creating new files based on the default template. By default, the new analysis 
folder will include all ref_period R files (.R, .Rmd, etc.) and their
containing folders, as well as .txt/.tex files.
}
\seealso{
Other functions for making template files/folders: 
\code{\link{new_project_individual}()},
\code{\link{new_project}()},
\code{\link{replace_params}()},
\code{\link{setup_data_dive}()},
\code{\link{setup_project}()}
}
\concept{functions for making template files/folders}
