% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{new_project}
\alias{new_project}
\title{Setup a new state project with default directories and template scripts}
\usage{
new_project(
  state,
  period,
  analysis_path = file.path("E:/SA/Projects/Data-Dashboards", state, period),
  sensitive_path = file.path("E:/SA/Data-Sensitive/Data-Dashboards", state,
    paste0("raw-", period)),
  production_path = file.path("E:/SA/Data-Production/Data-Dashboards", state),
  template_path = "template",
  print_message = "A new license data project has been initialized"
)
}
\arguments{
\item{state}{character: Two letter state designation}

\item{period}{character: Time period for the first project 
(e.g., "2019-q4", "2020-q4", etc.)}

\item{analysis_path}{character: File path for the analysis}

\item{sensitive_path}{character: File path for the sensitive data. If NULL,
no folders will be created.}

\item{production_path}{character: File path for the production data. If NULL, 
no folders will be created.}

\item{template_path}{character: Name of inst subpath in the package that points
to the template files}

\item{print_message}{character: Message to print on success}
}
\description{
This is intended to be run before data processing for a new state begins.
}
\examples{
# new_project("XX", "2019-q4")
}
\seealso{
Other functions for making template files/folders: 
\code{\link{new_project_individual}()},
\code{\link{replace_params}()},
\code{\link{setup_project}()},
\code{\link{update_project}()}
}
\concept{functions for making template files/folders}
