% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{replace_params}
\alias{replace_params}
\title{Helper function to replace parameter strings}
\usage{
replace_params(files, new, old)
}
\arguments{
\item{files}{full path to file(s) that containts strings to be replaced}

\item{new}{vector of new values to be used}

\item{old}{corresponding vector of old values to be replaced}
}
\description{
Intended only to be called from other lictemplate functions. Particularly for 
use in the "params.R" file.
}
\seealso{
Other functions for making template files/folders: 
\code{\link{new_project_individual}()},
\code{\link{new_project}()},
\code{\link{setup_data_dive}()},
\code{\link{setup_project}()},
\code{\link{update_project}()}
}
\concept{functions for making template files/folders}
